product list view model completed
