+++ src/User_Authentication/Models/ProductViewModels/ProductList.cs
@@ -1,0 +1,34 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using User_Authentication.Models;
+using User_Authentication.Data;
+using Microsoft.AspNetCore.Mvc.Rendering;
+
+namespace User_Authentication.Models.ProductViewModels
+{
+    public class ProductList
+    {
+        public List<Product> ProductList { get; set; }
+
+        public CreateProduct(ApplicationDbContext ctx)
+        {
+
+            this.ProductTypeId = ctx.ProductType
+                                    .OrderBy(l => l.Label)
+                                    .AsEnumerable()
+                                    .Select(li => new SelectListItem
+                                    {
+                                        Text = li.Label,
+                                        Value = li.ProductTypeId.ToString()
+                                    }).ToList();
+
+            this.ProductTypeId.Insert(0, new SelectListItem
+            {
+                Text = "Choose category...",
+                Value = "0"
+            });
+        }
+    }
+}
